Fix #5631: Rename `StdDelegatingSerializer` to `StdConvertingSerializer` (3.0->3.1)
