+++ release-notes/VERSION
@@ -5,7 +5,7 @@ Versions: 3.x (for earlier see VERSION-2.x)
 === Releases ===
 ------------------------------------------------------------------------
 
-3.1.0-rc1 (27-Jan-2026)
+3.1.0 (not yet released)
 
 #221: Support alternate radixes when writing numeric values as strings
  (requested by Christopher C)
@@ -163,6 +163,7 @@ Versions: 3.x (for earlier see VERSION-2.x)
  (implemented by @cowtowncoder, w/ Claude code)
 #5621: Fix typo: MissingInjectableValueExcepion → MissingInjectableValueException
  (implemented by Lee Jiwon)
+#5631: Rename `StdDelegatingSerializer` to `StdConvertingSerializer` (3.0->3.1)
 
 3.0.4 (not yet released)
